ml/pyproject.toml
[project]
name = "ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.3.1",
    "dotenv>=0.9.9",
    "neo4j>=6.1.0",
    "numpy>=2.4.2",
    "openai>=2.24.0",
    "pymupdf>=1.27.2",
    "python-docx>=1.2.0",
    "rich>=14.3.3",
    "sentence-transformers>=5.2.3",
]

[project.scripts]
sow-kg = "main:cli"
